frontend/requirements.txt
streamlit==1.35.0
sseclient-py==1.8.0
requests==2.31.0
